// src/openapi/types.gen.ts
export type CreateCompletion = {
  user: {
    id: string
  }
  tenant: {
    id: string
  }
  document:
    | CreateCompletionReferralLetterDocument
    | CreateCompletionProgressReportDocument
    | CreateCompletionNurseLogEntryDocument
  prompt:
    | CreateCompletionDocumentPrompt
    | CreateCompletionParagraphPrompt
    | CreateCompletionInlinePrompt
    | CreateCompletionCommentPrompt
}

export type CreateCompletionCommentPrompt = CreateCompletionPrompt & {
  type: "comment"
  draftContent: string
  commentContent: string
  highlightStartPosition: number
  highlightEndPosition: number
}

export type type = "comment"

export type CreateCompletionDocument = {
  id: string
}

export type CreateCompletionDocumentPrompt = CreateCompletionPrompt & {
  type: "document"
  content: string
}

export type type2 = "document"

export type CreateCompletionInlinePrompt = CreateCompletionPrompt & {
  type: "inline"
  content: string
  caretPosition: number
}

export type type3 = "inline"

export type CreateCompletionNurseLogEntryDocument = CreateCompletionDocument & {
  id: string
  type: "nurseLogEntry"
}

export type type4 = "nurseLogEntry"

export type CreateCompletionParagraphPrompt = CreateCompletionPrompt & {
  type: "paragraph"
  content: string
}

export type type5 = "paragraph"

export type CreateCompletionProgressReportDocument =
  CreateCompletionDocument & {
    id: string
    type: "progressReport"
    sectionType?: ProgressReportSectionType
  }

export type type6 = "progressReport"

export type CreateCompletionPrompt = {
  language?: Language
}

export type CreateCompletionReferralLetterDocument =
  CreateCompletionDocument & {
    id: string
    type: "referralLetter"
    sectionType?: ReferralLetterSectionType
  }

export type type7 = "referralLetter"

export type CreateSummary = {
  user: {
    id: string
  }
  tenant: {
    id: string
  }
  prompt: CreateSummarySimplePrompt
}

export type CreateSummarySimplePrompt = CreatSummaryPrompt & {
  type: "simple"
  content: string
}

export type type8 = "simple"

export type CreatSummaryPrompt = {
  language?: Language
}

export type DocumentType = "progressReport" | "referralLetter" | "nurseLogEntry"

export type Language = "de" | "fr" | "it" | "en"

export type ParameterapiVersionHeader = string

export type ParameteridPathParam = string

export type ProblemDetails = {
  status: number
  title: string
  detail: string
  [key: string]: unknown | number | string
}

export type ProgressReportSectionType =
  | "contactReason"
  | "anamnesis"
  | "currentCondition"
  | "progress"
  | "diagnosis"
  | "treatment"
  | "plan"
  | "prognosis"

export type PromptName = "general"

export type PromptTemplate = {
  name: PromptName
  version: string
  parameters: Array<string>
}

export type ReferralLetterSectionType =
  | "contactReason"
  | "anamnesis"
  | "currentCondition"
  | "progress"
  | "diagnosis"
  | "treatment"
  | "plan"
  | "prognosis"
// ...
export type CreateTranscriptionData = {
  body: {
    userId: string
    tenantId: string
    audioFile: Blob | File
  }
  headers?: {
    "X-Api-Version"?: string
  }
}